Move storage usage to plan page
