+++ src/views/pages/dashboard/plan/index.css
@@ -6,8 +6,7 @@
 
 .main {
   font-size: 1.1rem;
-  min-height: 300px;
-  padding: 10px;
+  padding: 10px 10px 30px 10px;
 }
 
 .header {
@@ -34,4 +33,16 @@
 
 .support-email {
   margin: 0 7px;
+}
+
+.usage {
+  font-weight: bold;
+  display: inline-block;
+  margin-right: 5px;
+}
+
+.total {
+  font-weight: bold;
+  margin-left: 5px;
+  display: inline-block;
 }
